+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - Dependencies now use the TOML syntax instead of a custom one [#39](https://github.com/wojciechkepka/pkger/pull/39)
 - Commands in configure, build and install scripts now use TOML syntax [#40](https://github.com/wojciechkepka/pkger/pull/40)
 - Add `--trace` option that sets log level of pkger to trace and make `--debug` actually set debug
+- Add some more fields for RPM builds, rename `section` to `group` and use it for RPM as well as DEB [#41](https://github.com/wojciechkepka/pkger/pull/41)